+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ## Master
 
+Move report logic on a separate file [#111](https://github.com/danger/danger-swift/pull/111) by [@f-meloni][]
+
 ## 0.7.0
 
 * Suggestions support [#110](https://github.com/danger/danger-swift/pull/110) by [@f-meloni][]