+++ CHANGELOG.md
@@ -13,9 +13,11 @@
 
 ## Master
 
+*Add help command to danger-swift [#122](https://github.com/danger/danger-swift/pull/122) by [@f-meloni][]
+
 ## 0.7.1
 
-Move report logic on a separate file [#111](https://github.com/danger/danger-swift/pull/111) by [@f-meloni][]
+*Move report logic on a separate file [#111](https://github.com/danger/danger-swift/pull/111) by [@f-meloni][]
 
 ## 0.7.0